Changed the URL the pages are downloaded from
Changed the URL the pages are downloaded from. By default, MangaFox
shows the page from its CDN but has also a "static" (and way faster)
site to get the images. Switched to the "static" way.

Hidden files (such as Thumbs.db on Windows) are ignored when counting a
chapter's pages count. It could sometimes lead to counting errors.